pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xblock-bunny"
version = "0.1.0"
description = "Open edX XBlock that embeds Bunny.net Stream videos with Token Authentication, TUS upload, and encoding-status webhooks."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "Apache-2.0" }
authors = [{ name = "Cubite - Amir Tadrisi", email = "amir@cubite.io" }]
keywords = ["openedx", "xblock", "bunny", "bunnycdn", "stream", "video"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Intended Audience :: Education",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.11",
    "Topic :: Education",
]
dependencies = [
    "XBlock>=1.10",
    "Django>=3.2",
    "djangorestframework>=3.14",
    "requests>=2.28",
    "cryptography>=41.0",
]

[project.urls]
Homepage = "https://github.com/amirtds/bunny-xblock"
Issues = "https://github.com/amirtds/bunny-xblock/issues"

[project.entry-points."xblock.v1"]
bunny_video = "bunny_xblock.xblock:BunnyVideoXBlock"

# Open edX auto-wires Django apps declared via these entry points
# (see edx_django_utils.plugins). Tutor / edx-platform discovers the
# package at boot — no INSTALLED_APPS edits required.
[project.entry-points."lms.djangoapp"]
xblock_bunny = "bunny_xblock.apps:BunnyXBlockAppConfig"

[project.entry-points."cms.djangoapp"]
xblock_bunny = "bunny_xblock.apps:BunnyXBlockAppConfig"

[tool.setuptools.packages.find]
include = ["bunny_xblock*"]

[tool.setuptools.package-data]
bunny_xblock = [
    "templates/bunny_xblock/*.html",
    "static/js/*.js",
    "static/js/vendor/*.js",
    "static/css/*.css",
]
